fix: SwiftDataModel의 프로퍼티, Extension 함수를 수정했습니다
